requirements.txt
# Spectral Analysis System - Requirements
#Python=3.14.2

# GUI Framework
PyQt6==6.10.1

# Data Processing
pandas>=1.5.0
numpy>=1.23.0
openpyxl>=3.0.0

# Plotting and Visualization
matplotlib>=3.6.0
pyqtgraph>=0.13.0

# Database
PyMySQL>=1.0.0

# Scientific Computing (for calibration module if needed)
scikit-learn>=1.2.0
scipy>=1.10.0
cryptography>=42.0.0
python-dotenv>=1.0.0

# to run file
# .\.venv311\Scripts\python.exe main.py
